API/requirements.txt
fastapi==0.100.0
uvicorn==0.21.0
requests==2.31.0
python-dotenv==1.0.0
openai==0.27.0
